Adjust tag tests
